local_vector_database/requirements.txt
langchain==0.2.14
langchain-community
langchain-core==0.2.33
langchain-text-splitters==0.2.2
langchain-openai
pypdf
faiss-cpu
langchainhub
langsmith
langgraph
langchain-pinecone
requests
black
openai
google-search-results
python-dotenv